[advanced] useValidation 훅 분리
